tools/ppt-export-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-pptx==1.0.2
pillow==11.2.1
